Remove JS sync client
