+++ packages/common/src/client/sync/bucket/BucketStorageAdapter.ts
@@ -1,34 +1,15 @@
 import { BaseListener, BaseObserverInterface, Disposable } from '../../../utils/BaseObserver.js';
 import { CrudBatch } from './CrudBatch.js';
-import { CrudEntry, OpId } from './CrudEntry.js';
-import { SyncDataBatch } from './SyncDataBatch.js';
+import { CrudEntry } from './CrudEntry.js';
 
 export interface BucketDescription {
   name: string;
   priority: number;
 }
 
-export interface Checkpoint {
-  last_op_id: OpId;
-  buckets: BucketChecksum[];
-  write_checkpoint?: string;
-  streams?: any[];
-}
-
 export interface BucketState {
   bucket: string;
   op_id: string;
-}
-
-export interface ChecksumCache {
-  checksums: Map<string, { checksum: BucketChecksum; last_op_id: OpId }>;
-  lastOpId: OpId;
-}
-
-export interface SyncLocalDatabaseResult {
-  ready: boolean;
-  checkpointValid: boolean;
-  checkpointFailures?: string[];
 }
 
 export type SavedProgress = {
@@ -37,26 +18,6 @@ export type SavedProgress = {
 };
 
 export type BucketOperationProgress = Record<string, SavedProgress>;
-
-export interface BucketChecksum {
-  bucket: string;
-  priority?: number;
-  /**
-   * 32-bit unsigned hash.
-   */
-  checksum: number;
-
-  /**
-   * Count of operations - informational only.
-   */
-  count?: number;
-  /**
-   * The JavaScript client does not use this field, which is why it's defined to be `any`. We rely on the structure of
-   * this interface to pass custom `BucketChecksum`s to the Rust client in unit tests, which so all fields need to be
-   * present.
-   */
-  subscriptions?: any;
-}
 
 export enum PSInternalTable {
   DATA = 'ps_data',
@@ -82,27 +43,14 @@ export interface BucketStorageListener extends BaseListener {
 
 export interface BucketStorageAdapter extends BaseObserverInterface<BucketStorageListener>, Disposable {
   init(): Promise<void>;
-  saveSyncData(batch: SyncDataBatch, fixedKeyFormat?: boolean): Promise<void>;
-  removeBuckets(buckets: string[]): Promise<void>;
-  setTargetCheckpoint(checkpoint: Checkpoint): Promise<void>;
 
-  startSession(): void;
-
-  getBucketStates(): Promise<BucketState[]>;
-  getBucketOperationProgress(): Promise<BucketOperationProgress>;
   hasMigratedSubkeys(): Promise<boolean>;
   migrateToFixedSubkeys(): Promise<void>;
-
-  syncLocalDatabase(
-    checkpoint: Checkpoint,
-    priority?: number
-  ): Promise<{ checkpointValid: boolean; ready: boolean; failures?: any[] }>;
 
   nextCrudItem(): Promise<CrudEntry | undefined>;
   hasCrud(): Promise<boolean>;
   getCrudBatch(limit?: number): Promise<CrudBatch | null>;
 
-  hasCompletedSync(): Promise<boolean>;
   updateLocalTarget(cb: () => Promise<string>): Promise<boolean>;
   getMaxOpId(): string;
