--- packages/common/src/client/sync/bucket/SyncDataBatch.ts
@@ -1,11 +1,0 @@
-import { SyncDataBucket } from './SyncDataBucket.js';
-
-// TODO JSON
-
-export class SyncDataBatch {
-  static fromJSON(json: any) {
-    return new SyncDataBatch(json.buckets.map((bucket: any) => SyncDataBucket.fromRow(bucket)));
-  }
-
-  constructor(public buckets: SyncDataBucket[]) {}
-}
