+++ packages/node/tests/sync.test.ts
@@ -3,9 +3,7 @@ import { beforeEach, describe, expect, vi } from 'vitest';
 
 import {
   AbstractPowerSyncDatabase,
-  BucketChecksum,
   createLogger,
-  OplogEntryJSON,
   PowerSyncConnectionOptions,
   ProgressWithOperations,
   Schema,
@@ -21,22 +19,13 @@ import {
   TestConnector,
   waitForSyncStatus
 } from './utils';
+import { BucketChecksum, OplogEntryJSON } from './sync_protocol';
 
 describe('Sync', () => {
-  function configureLegacyAndRustClient(bson: boolean) {
-    describe('js client', () => {
-      defineSyncTests(SyncClientImplementation.JAVASCRIPT, bson);
-    });
+  describe('json', () => defineSyncTests(false));
+  describe('bson', () => defineSyncTests(true));
 
-    describe('rust client', () => {
-      defineSyncTests(SyncClientImplementation.RUST, bson);
-    });
-  }
-
-  describe('json', () => configureLegacyAndRustClient(false));
-  describe('bson', () => configureLegacyAndRustClient(true));
-
-  mockSyncServiceTest('can migrate between sync implementations', async ({ syncService }) => {
+  mockSyncServiceTest.skip('can migrate between sync implementations', async ({ syncService }) => {
     function addData(id: string) {
       syncService.pushLine({
         data: {
@@ -137,9 +126,8 @@ describe('Sync', () => {
   });
 });
 
-function defineSyncTests(impl: SyncClientImplementation, bson: boolean) {
+function defineSyncTests(bson: boolean) {
   const options: PowerSyncConnectionOptions = {
-    clientImplementation: impl,
     connectionMethod: SyncStreamConnectionMethod.HTTP
   };
 
@@ -756,208 +744,154 @@ function defineSyncTests(impl: SyncClientImplementation, bson: boolean) {
     await another.waitForFirstSync({ priority: 0 });
   });
 
-  if (impl == SyncClientImplementation.RUST) {
-    mockSyncServiceTest('raw tables with inferred statements', async ({ syncService }) => {
-      const customSchema = new Schema({});
-      customSchema.withRawTables({
-        lists: {
-          schema: {}
-        }
-      });
-
-      const powersync = await syncService.createDatabase({ schema: customSchema });
-      await powersync.execute('CREATE TABLE lists (id TEXT NOT NULL PRIMARY KEY, name TEXT);');
-
-      const query = powersync.watchWithAsyncGenerator('SELECT * FROM lists')[Symbol.asyncIterator]();
-      expect((await query.next()).value.rows._array).toStrictEqual([]);
-
-      powersync.connect(new TestConnector(), options);
-      await vi.waitFor(() => expect(syncService.connectedListeners).toHaveLength(1));
-
-      syncService.pushLine({
-        checkpoint: {
-          last_op_id: '1',
-          buckets: [bucket('a', 1)]
-        }
-      });
-      syncService.pushLine({
-        data: {
-          bucket: 'a',
-          data: [
-            {
-              checksum: 0,
-              op_id: '1',
-              op: 'PUT',
-              object_id: 'my_list',
-              object_type: 'lists',
-              data: '{"name": "custom list"}'
-            }
-          ]
-        }
-      });
-      syncService.pushLine({ checkpoint_complete: { last_op_id: '1' } });
-      await powersync.waitForFirstSync();
-
-      expect((await query.next()).value.rows._array).toStrictEqual([{ id: 'my_list', name: 'custom list' }]);
-
-      syncService.pushLine({
-        checkpoint: {
-          last_op_id: '2',
-          buckets: [bucket('a', 2)]
-        }
-      });
-      await vi.waitFor(() => powersync.currentStatus.dataFlowStatus.downloading == true);
-      syncService.pushLine({
-        data: {
-          bucket: 'a',
-          data: [
-            {
-              checksum: 0,
-              op_id: '2',
-              op: 'REMOVE',
-              object_id: 'my_list',
-              object_type: 'lists'
-            }
-          ]
-        }
-      });
-      syncService.pushLine({ checkpoint_complete: { last_op_id: '2' } });
-      await vi.waitFor(() => powersync.currentStatus.dataFlowStatus.downloading == false);
-
-      expect((await query.next()).value.rows._array).toStrictEqual([]);
+  mockSyncServiceTest('raw tables with inferred statements', async ({ syncService }) => {
+    const customSchema = new Schema({});
+    customSchema.withRawTables({
+      lists: {
+        schema: {}
+      }
     });
 
-    mockSyncServiceTest('raw tables with explicit statements', async ({ syncService }) => {
-      const customSchema = new Schema({});
-      customSchema.withRawTables({
-        lists: {
-          put: {
-            sql: 'INSERT OR REPLACE INTO lists (id, name, _rest) VALUES (?, ?, ?)',
-            params: ['Id', { Column: 'name' }, 'Rest']
-          },
-          delete: {
-            sql: 'DELETE FROM lists WHERE id = ?',
-            params: ['Id']
+    const powersync = await syncService.createDatabase({ schema: customSchema });
+    await powersync.execute('CREATE TABLE lists (id TEXT NOT NULL PRIMARY KEY, name TEXT);');
+
+    const query = powersync.watchWithAsyncGenerator('SELECT * FROM lists')[Symbol.asyncIterator]();
+    expect((await query.next()).value.rows._array).toStrictEqual([]);
+
+    powersync.connect(new TestConnector(), options);
+    await vi.waitFor(() => expect(syncService.connectedListeners).toHaveLength(1));
+
+    syncService.pushLine({
+      checkpoint: {
+        last_op_id: '1',
+        buckets: [bucket('a', 1)]
+      }
+    });
+    syncService.pushLine({
+      data: {
+        bucket: 'a',
+        data: [
+          {
+            checksum: 0,
+            op_id: '1',
+            op: 'PUT',
+            object_id: 'my_list',
+            object_type: 'lists',
+            data: '{"name": "custom list"}'
           }
-        }
-      });
-
-      const powersync = await syncService.createDatabase({ schema: customSchema });
-      await powersync.execute('CREATE TABLE lists (id TEXT NOT NULL PRIMARY KEY, name TEXT, _rest TEXT);');
-
-      const query = powersync.watchWithAsyncGenerator('SELECT * FROM lists')[Symbol.asyncIterator]();
-      expect((await query.next()).value.rows._array).toStrictEqual([]);
-
-      powersync.connect(new TestConnector(), options);
-      await vi.waitFor(() => expect(syncService.connectedListeners).toHaveLength(1));
-
-      syncService.pushLine({
-        checkpoint: {
-          last_op_id: '1',
-          buckets: [bucket('a', 1)]
-        }
-      });
-      syncService.pushLine({
-        data: {
-          bucket: 'a',
-          data: [
-            {
-              checksum: 0,
-              op_id: '1',
-              op: 'PUT',
-              object_id: 'my_list',
-              object_type: 'lists',
-              data: '{"name": "custom list", "additional": "foo"}'
-            }
-          ]
-        }
-      });
-      syncService.pushLine({ checkpoint_complete: { last_op_id: '1' } });
-      await powersync.waitForFirstSync();
-
-      expect((await query.next()).value.rows._array).toStrictEqual([
-        { id: 'my_list', name: 'custom list', _rest: '{"additional":"foo"}' }
-      ]);
-
-      syncService.pushLine({
-        checkpoint: {
-          last_op_id: '2',
-          buckets: [bucket('a', 2)]
-        }
-      });
-      await vi.waitFor(() => powersync.currentStatus.dataFlowStatus.downloading == true);
-      syncService.pushLine({
-        data: {
-          bucket: 'a',
-          data: [
-            {
-              checksum: 0,
-              op_id: '2',
-              op: 'REMOVE',
-              object_id: 'my_list',
-              object_type: 'lists'
-            }
-          ]
-        }
-      });
-      syncService.pushLine({ checkpoint_complete: { last_op_id: '2' } });
-      await vi.waitFor(() => powersync.currentStatus.dataFlowStatus.downloading == false);
-
-      expect((await query.next()).value.rows._array).toStrictEqual([]);
+        ]
+      }
     });
-  } else {
-    mockSyncServiceTest('warns about raw tables', async ({ syncService }) => {
-      const customSchema = new Schema({});
-      customSchema.withRawTables({
-        lists: {
-          put: {
-            sql: 'INSERT OR REPLACE INTO lists (id, name) VALUES (?, ?)',
-            params: ['Id', { Column: 'name' }]
-          },
-          delete: {
-            sql: 'DELETE FROM lists WHERE id = ?',
-            params: ['Id']
+    syncService.pushLine({ checkpoint_complete: { last_op_id: '1' } });
+    await powersync.waitForFirstSync();
+
+    expect((await query.next()).value.rows._array).toStrictEqual([{ id: 'my_list', name: 'custom list' }]);
+
+    syncService.pushLine({
+      checkpoint: {
+        last_op_id: '2',
+        buckets: [bucket('a', 2)]
+      }
+    });
+    await vi.waitFor(() => powersync.currentStatus.dataFlowStatus.downloading == true);
+    syncService.pushLine({
+      data: {
+        bucket: 'a',
+        data: [
+          {
+            checksum: 0,
+            op_id: '2',
+            op: 'REMOVE',
+            object_id: 'my_list',
+            object_type: 'lists'
           }
+        ]
+      }
+    });
+    syncService.pushLine({ checkpoint_complete: { last_op_id: '2' } });
+    await vi.waitFor(() => powersync.currentStatus.dataFlowStatus.downloading == false);
+
+    expect((await query.next()).value.rows._array).toStrictEqual([]);
+  });
+
+  mockSyncServiceTest('raw tables with explicit statements', async ({ syncService }) => {
+    const customSchema = new Schema({});
+    customSchema.withRawTables({
+      lists: {
+        put: {
+          sql: 'INSERT OR REPLACE INTO lists (id, name, _rest) VALUES (?, ?, ?)',
+          params: ['Id', { Column: 'name' }, 'Rest']
+        },
+        delete: {
+          sql: 'DELETE FROM lists WHERE id = ?',
+          params: ['Id']
         }
-      });
-
-      const logger = createLogger('test', { logLevel: Logger.TRACE });
-      const logMessages: string[] = [];
-      (logger as any).invoke = (level, args) => {
-        console.log(...args);
-        logMessages.push(util.format(...args));
-      };
-
-      const powersync = await syncService.createDatabase({ schema: customSchema, logger });
-      powersync.connect(new TestConnector(), options);
-
-      await vi.waitFor(() => {
-        expect(logMessages).toEqual(
-          expect.arrayContaining([expect.stringContaining('Raw tables require the Rust-based sync client')])
-        );
-      });
+      }
     });
 
-    mockSyncServiceTest(`does not warn about raw tables if they're not used`, async ({ syncService }) => {
-      // Regression test for https://github.com/powersync-ja/powersync-js/issues/672
-      const customSchema = new Schema({});
+    const powersync = await syncService.createDatabase({ schema: customSchema });
+    await powersync.execute('CREATE TABLE lists (id TEXT NOT NULL PRIMARY KEY, name TEXT, _rest TEXT);');
 
-      const logger = createLogger('test', { logLevel: Logger.TRACE });
-      const logMessages: string[] = [];
-      (logger as any).invoke = (level, args) => {
-        console.log(...args);
-        logMessages.push(util.format(...args));
-      };
+    const query = powersync.watchWithAsyncGenerator('SELECT * FROM lists')[Symbol.asyncIterator]();
+    expect((await query.next()).value.rows._array).toStrictEqual([]);
 
-      const powersync = await syncService.createDatabase({ schema: customSchema, logger });
-      powersync.connect(new TestConnector(), options);
+    powersync.connect(new TestConnector(), options);
+    await vi.waitFor(() => expect(syncService.connectedListeners).toHaveLength(1));
 
-      await vi.waitFor(() => expect(syncService.connectedListeners).toHaveLength(1));
-      expect(logMessages).not.toEqual(
-        expect.arrayContaining([expect.stringContaining('Raw tables require the Rust-based sync client')])
-      );
+    syncService.pushLine({
+      checkpoint: {
+        last_op_id: '1',
+        buckets: [bucket('a', 1)]
+      }
     });
-  }
+    syncService.pushLine({
+      data: {
+        bucket: 'a',
+        data: [
+          {
+            checksum: 0,
+            op_id: '1',
+            op: 'PUT',
+            object_id: 'my_list',
+            object_type: 'lists',
+            data: '{"name": "custom list", "additional": "foo"}'
+          }
+        ]
+      }
+    });
+    syncService.pushLine({ checkpoint_complete: { last_op_id: '1' } });
+    await powersync.waitForFirstSync();
+
+    expect((await query.next()).value.rows._array).toStrictEqual([
+      { id: 'my_list', name: 'custom list', _rest: '{"additional":"foo"}' }
+    ]);
+
+    syncService.pushLine({
+      checkpoint: {
+        last_op_id: '2',
+        buckets: [bucket('a', 2)]
+      }
+    });
+    await vi.waitFor(() => powersync.currentStatus.dataFlowStatus.downloading == true);
+    syncService.pushLine({
+      data: {
+        bucket: 'a',
+        data: [
+          {
+            checksum: 0,
+            op_id: '2',
+            op: 'REMOVE',
+            object_id: 'my_list',
+            object_type: 'lists'
+          }
+        ]
+      }
+    });
+    syncService.pushLine({ checkpoint_complete: { last_op_id: '2' } });
+    await vi.waitFor(() => powersync.currentStatus.dataFlowStatus.downloading == false);
+
+    expect((await query.next()).value.rows._array).toStrictEqual([]);
+  });
 
   mockSyncServiceTest('can reconnect based on query changes', async ({ syncService }) => {
     // Test for https://discord.com/channels/1138230179878154300/1399340612435710034/1399340612435710034
